Minor bugfixes and account linking qa

+++ app/src/components/profile/DiscordConnection.tsx
@@ -1,6 +1,5 @@
 "use client"
 
-import { Session } from "next-auth"
 import Image from "next/image"
 import { toast } from "sonner"
 
@@ -8,31 +7,33 @@ import { Button } from "@/components/common/Button"
 import { syncPrivyUser } from "@/db/privy"
 import { useUser } from "@/hooks/useUser"
 import { useLinkAccount, usePrivy } from "@privy-io/react-auth"
-export const DiscordConnection = ({ session }: { session: Session }) => {
+
+export const DiscordConnection = ({ userId }: { userId: string }) => {
 
   const { user: privyUser, unlinkDiscord } = usePrivy()
-  const { user, invalidate: invalidateUser } = useUser({ id: session.user.id, enabled: !!session.user })
-
-  const username = user?.discord || privyUser?.discord?.username;
+  const { user, invalidate: invalidateUser } = useUser({ id: userId, enabled: true })
 
   const { linkDiscord } = useLinkAccount({
     onSuccess: async ({ user: updatedPrivyUser, linkMethod }) => {
       if (linkMethod === "discord") {
-        toast.promise(syncPrivyUser(updatedPrivyUser).then(() => invalidateUser()), {
-          loading: "Linking discord...",
-          success: "Discord linked successfully",
-          error: "Failed to link discord",
-        })
+        toast.promise(syncPrivyUser(updatedPrivyUser)
+          .then(() => invalidateUser()),
+          {
+            loading: "Linking discord...",
+            success: "Discord linked successfully",
+            error: "Failed to link discord",
+          })
       }
     }
   })
 
   const handleUnlinkDiscord = () => {
-    if (privyUser?.discord?.subject) {
+    if (user?.discord && privyUser?.discord?.subject) {
       toast.promise(unlinkDiscord(privyUser.discord.subject), {
         loading: "Unlinking discord...",
         success: (updatedPrivyUser) => {
-          syncPrivyUser(updatedPrivyUser).then(() => invalidateUser())
+          syncPrivyUser(updatedPrivyUser)
+            .then(() => invalidateUser())
           return "Discord unlinked successfully"
         },
         error: "Failed to unlink discord",
@@ -42,7 +43,7 @@ export const DiscordConnection = ({ session }: { session: Session }) => {
 
   return (
     <div className="flex flex-row gap-2">
-      {username && (
+      {user?.discord && (
         <div className="flex flex-col gap-2">
           <div>
             <div className="flex flex-1 p-3 border items-center gap-1.5 rounded-lg h-10">
@@ -52,14 +53,14 @@ export const DiscordConnection = ({ session }: { session: Session }) => {
                 width={16.67}
                 alt="Verified"
               />
-              <p className="text-sm">@{username}</p>
+              <p className="text-sm">@{user.discord}</p>
             </div>
           </div>
         </div>
       )}
 
 
-      {username ?
+      {user?.discord ?
         <Button variant="secondary" onClick={handleUnlinkDiscord}>Disconnect</Button>
         :
         <Button variant="primary" onClick={linkDiscord}>Connect</Button>
